Car_mpg_LR
Linear regression is a fundamental machine learning algorithm used to model the relationship between independent variables, such as engine characteristics, and a dependent variable, in this case, car MPG. By training a linear regression model on a dataset of cars and their MPG values, the algorithm learns to make predictions about the fuel efficiency of cars based on their features. This predictive model can help in understanding the impact of different car attributes on MPG and can be used for optimizing fuel efficiency, making informed purchasing decisions, or improving automotive designs. Linear regression is an essential tool in the field of machine learning for regression tasks, where the goal is to predict a continuous numeric outcome.